Renamed hot.h -> single_node_morton_hot.h
