+++ cpp/include/tapas/single_node_morton_hot.h
@@ -679,6 +679,35 @@ ProductIterator<CellIterator<hot::Cell<TSP>>,
         CellIterType(c1), CellIterType(c2));
 }
 
+class SingleNodeMortonHOT { };
+
+template <int DIM, class FP, class BT,
+          class BT_ATTR, class CELL_ATTR,
+          class PartitionAlgorithm>
+class Tapas;
+
+// Specialization of Tapas for HOT (single node Morton HOT) algorithm
+template <int DIM, class FP, class BT,
+          class BT_ATTR, class CELL_ATTR>
+class Tapas<DIM, FP, BT, BT_ATTR, CELL_ATTR, SingleNodeMortonHOT> {
+    typedef TapasStaticParams<DIM, FP, BT, BT_ATTR, CELL_ATTR> TSP; // Tapas static params
+  public:
+    typedef tapas::Vec<2, FP> Vec2;
+    typedef tapas::Vec<3, FP> Vec3;
+    typedef tapas::Region<TSP> Region;  
+    typedef tapas::hot::Cell<TSP> Cell;
+    //typedef tapas::BodyIterator<DIM, BT, BT_ATTR, Cell> BodyIterator;
+    typedef tapas::BodyIterator<Cell> BodyIterator;  
+    static Cell *Partition(typename BT::type *b,
+                           index_t nb, const Region &r,
+                           int max_nb) {
+        tapas::hot::Partition<TSP> part(max_nb);
+        return part(b, nb, r);
+    }
+};
+
+
+
 } // namespace tapas
 
 #endif // TAPAS_HOT_
